# Turn off debuginfo for dev, test, bench; turn on thinLTO for bench