--- a/Cargo.toml
+++ b/Cargo.toml
@@ -181,7 +181,7 @@ members = [
 
 [profile.dev]
 opt-level = 0
-debug = true
+debug = false
 codegen-units = 4
 lto = false
 incremental = true
@@ -204,7 +204,7 @@ rpath = false
 
 [profile.test]
 opt-level = 0
-debug = 2
+debug = false
 codegen-units = 16
 lto = false
 incremental = true
@@ -215,7 +215,7 @@ rpath = false
 # The benchmark profile is identical to release, except that lto = false
 [profile.bench]
 opt-level = 3
-debug = true
+debug = false
 codegen-units = 1
 lto = 'thin'
 incremental = false