--- a/README.md
+++ b/README.md
@@ -156,6 +156,13 @@ cargo test $TESTNAME
 
 Our CI systems automatically test all the pull requests, so making sure the full suite passes the test before creating your PR is not strictly required. **All merged PRs must have passed CI test.**
 
+Note that, to reduce compilation time, TiKV builds do not include debugging information by default. The easiest way to enable debuginfo is to procede build commands with `RUSTFLAGS=-Cdebuginfo=true`.
+
+```bash
+RUSTFLAGS=-Cdebuginfo=true make
+RUSTFLAGS=-Cdebuginfo=true cargo build
+```
+
 ### Getting the rest of the system working
 
 To get other components ([TiDB](https://github.com/pingcap/tidb) and [PD](https://github.com/pingcap/pd)) working, we suggest you follow the [development guide](https://github.com/pingcap/docs/blob/master/dev-guide/development.md), because you need the `pd-server` at least to work alongside `tikv-server` for integration level testing.