--- a/README.md
+++ b/README.md
@@ -156,11 +156,11 @@ cargo test $TESTNAME
 
 Our CI systems automatically test all the pull requests, so making sure the full suite passes the test before creating your PR is not strictly required. **All merged PRs must have passed CI test.**
 
-Note that, to reduce compilation time, TiKV builds do not include debugging information by default. The easiest way to enable debuginfo is to procede build commands with `RUSTFLAGS=-Cdebuginfo=true`.
+Note that, to reduce compilation time, TiKV builds do not include debugging information by default. The easiest way to enable debuginfo is to procede build commands with `RUSTFLAGS=-Cdebuginfo=1` (for line numbers), or `RUSTFLAGS=-Cdebuginfo=2` (for full debuginfo).
 
 ```bash
-RUSTFLAGS=-Cdebuginfo=true make
-RUSTFLAGS=-Cdebuginfo=true cargo build
+RUSTFLAGS=-Cdebuginfo=2 make
+RUSTFLAGS=-Cdebuginfo=2 cargo build
 ```
 
 ### Getting the rest of the system working