--- a/scripts/run-cargo.sh
+++ b/scripts/run-cargo.sh
@@ -28,6 +28,9 @@ else
     exit 1
 fi
 
+# All features are enabled explicitly
+args="$args --no-default-features"
+
 features="default"
 
 if [[ -n "$X_CARGO_FEATURES" ]]; then