## What have you changed? (mandatory)

This is another compile-time improvement PR. It is moving a variety of cargo profile changes from the testing profiles in `etc/` to the actual profiles used by cargo and the makefile.

This turns debuginfo completely off for the dev, test, and bench profiles. Backtraces will continue to display symbols, but not line numbers, and the binaries are not debuggable.

Turning debuginfo off is the biggest compile-time win I've identified. This PR though does _not_ do so for the release profile, because I am first [splitting that into two different profilfes](https://github.com/tikv/tikv/pull/4687), one for devs, and one for building the distributable packages.

I've updated the README to indicate how to turn debuginfo back on, for both makefile and cargo use cases, but I really want that info to live in a [developer guide](https://github.com/tikv/tikv/issues/4673), and will move it there later.

Here are some old timings removing debuginfo for release builds:

```
before:

real    27m24.772s
user    62m0.855s
sys     2m22.194s

after:

real    21m39.722s
user    49m12.515s
sys     1m44.381s
```

The results I think should be proportionally better for dev builds because there less time is spent optimizing, but similar time is spent generating debuginfo, but I don't seem to have any dev timings in my notes that just test debuginfo removal. I _should_ really have those numbers, so I'll produce them in a few days.

And of course the binary sizes shrink dramatically.

This also lists _all_ profile options directly in the manifest, to make it perfectly clear what the possibilities are and what their values are. Note that the test profile is set to code-units=16, whereas the existing explicit code-units are either 1 or 4. The cargo manual claims 16 is the default for test, so that is what I've written. It suggests we could bump the dev profile to 16.

This also turns _on_ thinLTO for the bench profile. This will make `cargo bench` a bit slower, but I think is the right thing to do. Benchmarking without any LTO seems questionable.

## What are the type of the changes? (mandatory)

- Engineering (engineering change which doesn't change any feature or fix any issue)

## How has this PR been tested? (mandatory)

Various benchmarking. `cargo build -v`.

## Does this PR affect documentation (docs) or release note? (mandatory)

No

## Does this PR affect tidb-ansible update? (mandatory)

No

## Refer to a related PR or issue link (optional)

https://github.com/tikv/tikv/issues/4674

## Benchmark result if necessary (optional)

## Add a few positive/negative examples (optional)