@brson 2019-05-13
The method I've given for turning debuginfo on is ugly: `RUSTFLAGS=-Cdebuginfo=2 cargo build`. I don't know a simpler way to do it.

There are unstable cargo flags that will do it a little more cleanly, but they require passing extra flags to enable unstable features, so the overall solution is not simpler.

If anybody knows a simpler way to do it please tell me.

When I add it to the pending developer guide I will also mention `.cargo/config`, since it can be done there, but still not simpler than `RUSTFLAGS`.