@siddontang 2019-05-14
@cwen0 

can we benchmark this PR directly in our BenchBot platform?