@siddontang 2019-05-15
Hi @brson 

I remember that even we set `debug = false`, the total binary size is still huge, maybe we can re-check this.