@mahjonp 2019-05-16
/run-all-tests