@brson 2019-05-16
> I remember that even we set `debug = false`, the total binary size is still huge, maybe we can re-check this.

@siddontang 

Wow, you are right (for the dev profile). I'm shocked. Here's both dev and release with debuginfo=0. Release profile does as expected and emptys most of the debug sections. I cannot explain what's going on in the dev profile - cargo is passing debuginfo=0 (when it's set to 0 - whe it's set to "false" cargo passes nothing).

This seems like an obvious bug in rustc.

For these two tests I additionally set `debuginfo=0` for release profile, which is not part of this patch.

```
brian@DESKTOP-UCV672I:~/pingcap/tikv$ ls -lh target/*/tikv-server
-rwxrwxrwx 2 brian brian 187M May 16 15:38 target/debug/tikv-server
-rwxrwxrwx 2 brian brian  41M May 15 22:48 target/release/tikv-server
```

And bloaty:

```
brian@DESKTOP-UCV672I:~/pingcap/tikv$ bloaty target/debug/tikv-server
     VM SIZE                         FILE SIZE
 --------------                   --------------
   0.0%       0 .debug_info        61.5Mi  33.0%
  60.9%  46.6Mi .text              46.6Mi  25.1%
   0.0%       0 .strtab            18.3Mi   9.8%
   0.0%       0 .debug_str         14.1Mi   7.6%
   0.0%       0 .symtab            6.71Mi   3.6%
   8.0%  6.15Mi .dynstr            6.15Mi   3.3%
   7.7%  5.88Mi .eh_frame          5.88Mi   3.2%
   0.0%       0 .debug_line        5.44Mi   2.9%
   5.9%  4.51Mi .rela.dyn          4.51Mi   2.4%
   3.2%  2.45Mi .data.rel.ro       2.45Mi   1.3%
   0.0%       0 .debug_ranges      2.28Mi   1.2%
   2.8%  2.18Mi .bss                    0   0.0%
   2.5%  1.94Mi .rodata            1.94Mi   1.0%
   2.5%  1.94Mi .gcc_except_table  1.94Mi   1.0%
   2.5%  1.91Mi .dynsym            1.91Mi   1.0%
   0.0%       0 .debug_aranges     1.90Mi   1.0%
   2.1%  1.61Mi .eh_frame_hdr      1.61Mi   0.9%
   0.0%       0 .debug_abbrev      1.42Mi   0.8%
   0.9%   723Ki .got                723Ki   0.4%
   0.7%   516Ki .gnu.hash           516Ki   0.3%
   0.3%   199Ki [29 Others]         375Ki   0.2%
 100.0%  76.6Mi TOTAL               186Mi 100.0%
brian@DESKTOP-UCV672I:~/pingcap/tikv$ bloaty target/release/tikv-server
     VM SIZE                         FILE SIZE
 --------------                   --------------
  64.1%  17.4Mi .text              17.4Mi  42.7%
   0.0%       0 .debug_info        5.47Mi  13.4%
   0.0%       0 .debug_loc         4.34Mi  10.6%
   8.3%  2.25Mi .data.rel.ro       2.25Mi   5.5%
   0.0%       0 .strtab            2.18Mi   5.3%
   7.8%  2.13Mi .bss                    0   0.0%
   7.7%  2.09Mi .rela.dyn          2.09Mi   5.1%
   4.9%  1.34Mi .rodata            1.34Mi   3.3%
   4.6%  1.25Mi .eh_frame          1.25Mi   3.1%
   0.0%       0 .debug_ranges       948Ki   2.3%
   0.0%       0 .debug_line         859Ki   2.1%
   0.0%       0 .symtab             847Ki   2.0%
   0.0%       0 .debug_str          762Ki   1.8%
   1.5%   406Ki .gcc_except_table   406Ki   1.0%
   0.0%       0 .debug_abbrev       277Ki   0.7%
   0.7%   189Ki .eh_frame_hdr       189Ki   0.5%
   0.0%       0 .debug_macro       95.7Ki   0.2%
   0.0%       0 .debug_aranges     73.4Ki   0.2%
   0.2%  52.7Ki [29 Others]        62.4Ki   0.1%
   0.1%  36.1Ki .dynstr            36.1Ki   0.1%
   0.1%  24.4Ki .dynsym            24.4Ki   0.1%
 100.0%  27.2Mi TOTAL              40.8Mi 100.0%

```
Here are my previous, more detailed, tests, with the 'before' info:

```
before:

    Finished release [optimized + debuginfo] target(s) in 27m 24s

real    27m24.772s
user    62m0.855s
sys     2m22.194s

ubuntu@ip-172-30-0-238:~/tikv$ ls -l target/release/tikv-server
-rwxrwxr-x 2 ubuntu ubuntu 311546456 Feb 16 00:54 target/release/tikv-server

ubuntu@ip-172-30-0-238:~/tikv$ ../bloaty/bloaty target/release/tikv-server
     VM SIZE                         FILE SIZE
 --------------                   --------------
   0.0%       0 .debug_info         116Mi  39.4%
   0.0%       0 .debug_loc         66.8Mi  22.5%
   0.0%       0 .debug_str         32.0Mi  10.8%
   0.0%       0 .debug_ranges      22.3Mi   7.5%
  64.5%  18.6Mi .text              18.6Mi   6.3%
   0.0%       0 .debug_line        11.6Mi   3.9%
   0.0%       0 .debug_pubnames    7.57Mi   2.5%
   0.0%       0 .debug_pubtypes    7.34Mi   2.5%
   0.0%       0 .strtab            2.72Mi   0.9%
   8.0%  2.31Mi .data.rel.ro       2.31Mi   0.8%
   7.6%  2.21Mi .rela.dyn          2.21Mi   0.7%
   7.4%  2.14Mi .bss                    0   0.0%
   0.0%       0 .debug_abbrev      1.77Mi   0.6%
   5.0%  1.43Mi .rodata            1.43Mi   0.5%
   4.7%  1.37Mi .eh_frame          1.37Mi   0.5%
   0.0%       0 .symtab            1006Ki   0.3%
   1.7%   507Ki .gcc_except_table   507Ki   0.2%
   0.0%       0 .debug_aranges      322Ki   0.1%
   0.8%   227Ki .eh_frame_hdr       227Ki   0.1%
   0.0%       0 .debug_macro        111Ki   0.0%
   0.2%  62.3Ki [31 Others]        71.2Ki   0.0%
 100.0%  28.9Mi TOTAL               297Mi 100.0%

I'll try the same test with the "test" and "bench" profiles, then look into what's going on in the dev profile.

> can we benchmark this PR directly in our BenchBot platform?

Sure, but I do not know how. Who should I talk to?

> Without line number it's hard to identify which line goes wrong when a case fails.

@BusyJay good point. I've set debuginfo=1 for the test profile.