@brson 2019-05-16
@siddontang after some testing debuginfo=0 does reduce the size of the debug sections _some_ in dev mode, but not dramatically. I'm still trying to understand what the reasoning for that is.