@brson 2019-05-17
It's definitely rocksdb that is retaining all that debuginfo in the dev profile. I have been looking at the cmake crate to figure out if it's possible to shut off debuginfo in debug mode, but it's not obvious. It _seems_ like cmake wants  debuginfo on for its debug profile. Requires further investigation.

I think we should land this and file yet another bug to fix librocksdb_sys to work with debug mode + no debuginfo