@brson 2019-05-17
Here's the bug for getting rocksdb to turn off debuginfo https://github.com/tikv/tikv/issues/4711.