@brson 2019-05-17
So that 187MB binary in dev mode is seemingly _mostly_ C++ debuginfo.