@siddontang 2019-05-17
@brson 

Seem that if we can turn off debug info in RocksDB, we can again reduce the compiling time, right?