@brson 2019-05-17
I ended up hacking on cmake-rs for a while to try to solve the rocksdb debuginfo, but didn't quite get to a solution before becoming too tired to continue. I did file a bug against cmake-rs, which contains a link to my branch: https://github.com/alexcrichton/cmake-rs/issues/79