@brson 2019-05-17
@siddontang yes, but only for `dev` and `test` builds.